+++ ts/components/GameWrapper.tsx
@@ -188,6 +188,7 @@ export default class GameWrapper extends React.Component<
 							)
 						}
 						cards={this.props.cards}
+						cardsByDbfId={this.props.cardsByDbfId}
 						cardOracle={this.props.cardOracle}
 						mulliganOracle={this.props.mulliganOracle}
 						descriptors={this.props.state.descriptors}